requirements.txt
numpy
pandas
rasterio
rioxarray
xarray
scikit-learn
matplotlib
streamlit
pyproj
shapely
Flask
flask-cors
requests
python-dotenv
